requirements.txt
Flask
requests
tflite-runtime
Pillow
numpy
